PYIC-6738: Create Vtr class
