+++ lambdas/evaluate-gpg45-scores/src/main/java/uk/gov/di/ipv/core/evaluategpg45scores/EvaluateGpg45ScoresHandler.java
@@ -177,7 +177,7 @@ public class EvaluateGpg45ScoresHandler
             var gpg45Scores = gpg45ProfileEvaluator.buildScore(vcs);
 
             var requestedVotsByStrength =
-                    clientOAuthSessionItem.getRequestedVotsByStrengthDescending();
+                    clientOAuthSessionItem.getParsedVtr().getRequestedVotsByStrengthDescending();
             var supportedGpg45ProfilesByVotStrength =
                     requestedVotsByStrength.stream()
                             .filter(vot -> vot.getSupportedGpg45Profiles() != null)
