+++ lambdas/process-cri-callback/src/main/java/uk/gov/di/ipv/core/processcricallback/service/CriCheckingService.java
@@ -234,7 +234,9 @@ public class CriCheckingService {
             // Check CI levels against the lowest confidence identity requested so we don't send the
             // user on an unnecessary mitigation journey.
             var lowestConfidenceRequested =
-                    clientOAuthSessionItem.getLowestStrengthRequestedGpg45Vot(configService);
+                    clientOAuthSessionItem
+                            .getParsedVtr()
+                            .getLowestStrengthRequestedGpg45Vot(configService);
             var journeyResponse =
                     ciMitUtilityService.getMitigationJourneyIfBreaching(
                             cis, lowestConfidenceRequested);
