+++ libs/common-services/src/main/java/uk/gov/di/ipv/core/library/persistence/item/ClientOAuthSessionItem.java
@@ -7,14 +7,10 @@ import lombok.NoArgsConstructor;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbBean;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbPartitionKey;
 import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
-import uk.gov.di.ipv.core.library.domain.ProfileType;
-import uk.gov.di.ipv.core.library.enums.Vot;
-import uk.gov.di.ipv.core.library.service.ConfigService;
+import uk.gov.di.ipv.core.library.domain.Vtr;
 
 import java.util.Arrays;
 import java.util.List;
-
-import static uk.gov.di.ipv.core.library.config.CoreFeatureFlag.P1_JOURNEYS_ENABLED;
 
 @DynamoDbBean
 @ExcludeFromGeneratedCoverageReport
@@ -45,46 +41,8 @@ public class ClientOAuthSessionItem implements DynamodbItem {
         return Arrays.asList(this.scope.split(" "));
     }
 
-    public List<Vot> getRequestedVotsByStrengthDescending() {
-        return Vot.SUPPORTED_VOTS_BY_DESCENDING_STRENGTH.stream()
-                .filter(vot -> vtr.contains(vot.name()))
-                .toList();
-    }
-
-    // Refactor this out in PYIC-6984
-    public Vot getLowestStrengthRequestedGpg45Vot(ConfigService configService) {
-        var requestedGpg45VotsByStrengthDescending =
-                getRequestedVotsByStrengthDescending().stream()
-                        .filter(vot -> vot.getProfileType() == ProfileType.GPG45)
-                        .toList();
-
-        var lowestStrengthRequestedGpg45Vot =
-                requestedGpg45VotsByStrengthDescending.get(
-                        requestedGpg45VotsByStrengthDescending.size() - 1);
-
-        if (lowestStrengthRequestedGpg45Vot == Vot.P1
-                && !configService.enabled(P1_JOURNEYS_ENABLED)) {
-            lowestStrengthRequestedGpg45Vot =
-                    requestedGpg45VotsByStrengthDescending.get(
-                            requestedGpg45VotsByStrengthDescending.size() - 2);
-        }
-
-        return lowestStrengthRequestedGpg45Vot;
-    }
-
-    // Refactor this out in PYIC-6984
-    public Vot getLowestStrengthRequestedVot(ConfigService configService) {
-        var requestedVotsByStrengthDescending = getRequestedVotsByStrengthDescending();
-
-        var lowestStrengthRequestedVot =
-                requestedVotsByStrengthDescending.get(requestedVotsByStrengthDescending.size() - 1);
-
-        if (lowestStrengthRequestedVot == Vot.P1 && !configService.enabled(P1_JOURNEYS_ENABLED)) {
-            lowestStrengthRequestedVot =
-                    requestedVotsByStrengthDescending.get(
-                            requestedVotsByStrengthDescending.size() - 2);
-        }
-
-        return lowestStrengthRequestedVot;
+    // PYIC-6984 can we store the Vtr object directly in dynamo instead of creating it here?
+    public Vtr getParsedVtr() {
+        return new Vtr(vtr);
     }
 }
